"Removed unnecessary words in printf statement in main function of QueuesUsingArray.c; Binary of QueuesUsingArray.exe updated"
The diff shows that there is a change in the printf statement in the main function of QueuesUsingArray.c where the word "if you want to" is removed from "Type 1 if you want to dequeue and 0 if you want to exit:". Additionally, there is a change in the printf statement that displays the dequeued value where the words "was dequeued from the queue." is changed to "dequeued." Lastly, there is a change in the binary file of QueuesUsingArray.exe.